requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
Django==3.1.4
djangorestframework==3.12.2
flake8==3.8.4
flake8-django==1.1.1
jwt==1.1.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
python-environ==0.4.54
pytz==2020.4
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
